Make minified JSON
