--- CTFd/utils/exports/encoders.py
@@ -1,11 +1,0 @@
-import json
-from datetime import datetime, date
-from decimal import Decimal
-
-
-class JSONEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, (datetime, date)):
-            return obj.isoformat()
-        if isinstance(obj, Decimal):
-            return str(obj)
